Add skeleton loaders to dao feed
